% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcanR.R
\name{getCansimTbl}
\alias{getCansimTbl}
\title{statcanR}
\usage{
getCansimTbl(tableNumber, lang)
}
\arguments{
\item{tableNumber}{The table number of the Statistics Canada data table}

\item{lang}{The language wanted}
}
\value{
The output will be a data table representing the data associated with the chosen table number
}
\description{
Get all Canadian statistics data (CANSIM tables) now identify by Product IDs (PID) by the new Statistics Canada Web Data Service
}
\details{
The
\code{_getCansimTbl_} function has 2 arguments to fulfill to get data: {_tableNumber_} & param{_lang_}

The _tableNumber_ argument simply referes to the table number of the Statistics Canada data table you want to collect,
such as '_14-10-0287-03_' for the _Labour force characteristics by province, monthly, seasonally adjusted_ as an example.

To get the table number :\url{https://www150.statcan.gc.ca/n1/en/type/data}

The second argument, _lang_, referes to the language. As Canada is a billingual country (french speaking in Quebec and english speaking in the other provinces), Statistics Canada display all Statistics Data in both languages.
Therefore, users can choose if they want to get satistics data table in french or english by seting the _lang_ argument by c("fra", "eng").
}
\examples{
\dontrun{
datatable <- getCansimTble("14-10-0287-03","eng")
}
}
