% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statCanR.R
\name{statcan_download_data}
\alias{statcan_download_data}
\title{statcanR download data function}
\usage{
statcan_download_data(tableNumber, lang)
}
\arguments{
\item{tableNumber}{The table number of the Statistics Canada data table}

\item{lang}{The language wanted}
}
\value{
The output will be  a data table and csv file representing the data associated with the chosen table number.
}
\description{
statcanR download data function
}
\examples{
mydata <- statcan_data('27-10-0014-01', 'eng')
}
